+++ pages/index.php
@@ -1,4 +1,4 @@
 <?php
 
-echo rex_view::title(rex_i18n::msg('buchungskalender'));
+echo rex_view::title(rex_i18n::msg('buka_booking_calendar'));
 rex_be_controller::includeCurrentPageSubPath();